==> Code/validation/4 SH_examine_comparisons.Rmd <==
---
title: "Compare findings"
output: html_document
---

In this file we compare the LE estimates for Sam's Health Affairs paper to estimates we get from running our model on his data. 
Thus the goal of this exercise is to answer the questions "are the models doing the same thing?"

```{r load_libraries}
library(dplyr)
library(stringr)
library(ggplot2)
```

```{r read_data}
le.table <- read.csv("/Users/corinneriddell/Documents/repos/BlackWhiteMortalityGap/Code/validation/le_table.csv", header = T) 
#le_table was created in the SH_comparison.R file

SH.dat <- read.csv("/Users/corinneriddell/Documents/repos/BlackWhiteMortalityGap/Data/SH_HA_supp_app_data.csv",
                   header = T) #supplementary appendix data -- so Sam's original estimates
```

```{r tidy_data}
SH.dat <- SH.dat %>% rename(state = State, 
                            le.black.original = le.black,
                            le.white.original = le.white,
                            le.gap.original = le.gap, 
                            le.gap.ci.original = le.gap.ci)

compare.le <- merge(le.table, SH.dat, by = c("state", "sex", "year"))

test <- strsplit(stringr::str_sub(compare.le$le.gap.ci.original, 2, -2), ",")
test2 <- matrix(unlist(test), ncol=2, byrow=TRUE)

compare.le$le.gap.lcl.original <- as.numeric(test2[ , 1])
compare.le$le.gap.ucl.original <- as.numeric(test2[ , 2])
rm(test, test2)

compare.le <- compare.le %>% mutate(le.gap.width.original = le.gap.ucl.original - le.gap.lcl.original,
                                    le.gap.width = le.gap.ucl - le.gap.lcl,
                                    small.black.state = state %in% c("Hawaii", "Alaska", "Maine", 
                                                                     "New Hampshire","Utah", "New Mexico")
                                    )

write.csv(compare.le, "/Users/corinneriddell/Documents/repos/BlackWhiteMortalityGap/Code/validation/compare_LE.csv")
```

**Figure 1: Comparison of life expectancy between the original model and our model**

```{r}
ggplot(subset(compare.le, state!= "Wyoming" & year != "difference"), 
       aes(x = le.white.original, y = le.white)) +
  geom_point(aes(x = le.black.original, y = le.black, col = "blacks")) + 
  geom_point(aes(col = "whites")) +
  facet_grid(year~sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Life expectancy (years)") +
  xlab("Original life expectancy (years)") +
  geom_text(data = subset(compare.le, small.black.state == T & year != "difference"),
            aes(x = le.black.original, y = le.black, label = state), size = 2.5)
```

**Figure 2: Comparison of life expectancy change over time (1990 vs. 2009) between the original model and our model**

```{r}
ggplot(subset(compare.le, state!= "Wyoming" & year == "difference"), 
       aes(x = le.white.original, y = le.white)) +
  geom_point(aes(x = le.black.original, y = le.black, col = "blacks")) + 
  geom_point(aes(col = "whites")) +
  facet_grid(~sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Life expectancy change (years)") +
  xlab("Original life expectancy change (years)") +
  geom_text(data = subset(compare.le, small.black.state == T & year == "difference"),
            aes(x = le.black.original, y = le.black, label = state), size = 2.5)

```

**Figure 3: Comparison of life expectancy gap between the original model and our model**

```{r}
ggplot(subset(compare.le, state!= "Wyoming" & year != "difference"), 
       aes(x = le.gap.original, y = le.gap)) +
  geom_point() +
  facet_grid(year~sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Life expectancy gap (years)") +
  xlab("Original life expectancy gap (years)") +
  geom_text(data = subset(compare.le, small.black.state == T & year != "difference"),
            aes(x = le.gap.original, y = le.gap, label = state), size = 2.5)
```

**Figure 4: Comparison of change over time (1990 vs. 2009) in the life expectancy gap between the original model and our model**

```{r}
ggplot(subset(compare.le, state!= "Wyoming" & year == "difference"), 
       aes(x = le.gap.original, y = le.gap)) +
  geom_point() +
  facet_grid(.~sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Change in life expectancy gap (years)") +
  xlab("Original change in life expectancy gap (years)") +
  geom_text(data = subset(compare.le, small.black.state == T & year == "difference"),
            aes(x = le.gap.original, y = le.gap, label = state), size = 2.5)
```

**Figure 5: Comparison of CI width for LE gap between the original model and our model**

```{r}
ggplot(subset(compare.le, state!= "Wyoming" & year != "difference"), 
       aes(x = le.gap.width.original, y = le.gap.width)) +
  geom_point() +
  facet_grid(year~sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Width of CI for LE gap") +
  xlab("Original width of CI for LE gap") +
  geom_text(data = subset(compare.le, small.black.state == T & year != "difference"),
            aes(x = le.gap.width.original, y = le.gap.width, label = state), size = 2.5)
```

**Figure 6: Comparison of the change over time (1990 vs. 2009) in the CI width for LE gap between the original model and our model**

```{r}
ggplot(subset(compare.le, state!= "Wyoming" & year == "difference"), 
       aes(x = le.gap.width.original, y = le.gap.width)) +
  geom_point() +
  facet_grid(.~sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Width of CI for LE gap") +
  xlab("Original width of CI for LE gap") +
  geom_text(data = subset(compare.le, small.black.state == T & year == "difference"),
            aes(x = le.gap.width.original, y = le.gap.width, label = state), size = 2.5)
```
