==> Code/validation/5 SH_second_comparison.Rmd <==
---
title: "Compare findings"
output: html_document
---

In this file we compare our findings (using our data which is different because higher rate of missingness and more years of data) to Sam's Health Affairs appendix data. 

```{r load_libraries}
library(dplyr)
library(stringr)
library(ggplot2)
library(plotly)
```

```{r read_data}
BlackWhite_results <- read.csv("~/Documents/repos/BlackWhiteMortalityGap/Results2/BlackWhite_results.csv")
subset <- filter(BlackWhite_results, year %in% c(1990, 2009))
subset <- subset %>% select(state, year, sex, Census_Division, Census_Region, stabbrs, starts_with("LE"))

SH.dat <- read.csv("~/Documents/repos/BlackWhiteMortalityGap/Data/SH_HA_supp_app_data.csv", header = T)
```

```{r tidy_data}
SH.dat <- SH.dat %>% rename(state = State, 
                            le.black.SH = le.black,
                            le.white.SH = le.white,
                            le.gap.SH = le.gap, 
                            le.gap.ci.SH = le.gap.ci) %>%
  filter(year %in% c(1990, 2009))

compare.le <- merge(subset, SH.dat, by = c("state", "sex", "year"))

test <- strsplit(stringr::str_sub(compare.le$le.gap.ci.SH, 2, -2), ",")
test2 <- matrix(unlist(test), ncol=2, byrow=TRUE)

compare.le$le.gap.lcl.SH <- as.numeric(test2[ , 1])
compare.le$le.gap.ucl.SH <- as.numeric(test2[ , 2])
rm(test, test2)

compare.le <- compare.le %>% mutate(le.gap.width.SH = le.gap.ucl.SH - le.gap.lcl.SH,
                                    le.gap.width = LE_wbgap_ucl - LE_wbgap_lcl,
                                    small.black.state = state %in% c("Hawaii", "Alaska", "Maine", 
                                                                     "New Hampshire","Utah", "New Mexico")
                                    )
```

**Figure 1: Comparison of life expectancy between the original model and our model for whites**

```{r}
ggplotly(ggplot(compare.le, 
       aes(x = le.white.SH, y = LE_white_mean)) +
  geom_point(aes(col = state)) +
  facet_grid(year ~ sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Life expectancy mean (years)") +
  xlab("Life expectancy estimate from Health Affairs paper (years)"))
```

**Figure 2: Comparison of life expectancy between the original model and our model for blacks**

```{r}
ggplotly(ggplot(compare.le, 
       aes(x = le.black.SH, y = LE_black_mean)) +
  geom_point(aes(col = state)) +
  facet_grid(year ~ sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Life expectancy mean (years)") +
  xlab("Life expectancy estimate from Health Affairs paper (years)"))
```

**Figure 3: Comparison of life expectancy gap between the original model and our model**

```{r}
ggplotly(ggplot(compare.le, 
       aes(x = le.gap.SH, y = LE_wbgap_mean)) +
  geom_point(aes(col = state)) +
  facet_grid(year ~ sex ) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Life expectancy gap (years)") +
  xlab("Life expectancy gap from Health Affairs paper (years)")) 
```

**Figure 4: Comparison of CI width for LE gap between the original model and our model**

```{r}
ggplotly(ggplot(compare.le, 
       aes(x = le.gap.width.SH, y = le.gap.width)) +
  geom_point(aes(col = state)) +
  facet_grid(year ~ sex) +
  geom_abline(intercept = 0, slope = 1) + 
  ylab("Width of CI for LE gap") +
  xlab("Width of CI for LE gap from Health Affairs paper") )
```
